aeq.error: Fix wrong function name
`arguments` was pointing to the arguments of aeq.error, so the function name was always
anonymous. You can now pass the  object into aeq.error and get function name
and aruments passed to the function. If you do not pass in `arguments`, the function
will try to find the name of the function where the error was thrown, but the method
may not work 100% of the time.