DOM: Add deep argument to getItems
Previously the function only returned the topmost items in a folder, this
was different from how the function behaves when no argument is given,
where all items in the project is returned.

If no arg is given, it returns all items in the project.
If the folder is not found (when the argument is a string) an empty
arrayEx is returned.
If deep is not disabled, it returns all items in subfolders as well,
else it returns only the items in the folder.